deleted a cooldown,done some fixes

+++ include/retDamageScoreByTypeOfAttack.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "Constants and Types.h"
-int retDamageScoreByTypeOfAttack(Type::AttackType::Player_Attacks a);
-int retDamageScoreByTypeOfAttack(Type::AttackType::Magical_Monster_Attacks m);
-int retDamageScoreByTypeOfAttack(Type::AttackType::Physical_Monster_Attacks ph);
+short retDamageScoreByTypeOfAttack(Type::AttackType::Player_Attacks a);
+short retDamageScoreByTypeOfAttack(Type::AttackType::Magical_Monster_Attacks m);
+short retDamageScoreByTypeOfAttack(Type::AttackType::Physical_Monster_Attacks ph);
